@@ -1,59 +1,45 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 buildscript {
     ext.versions = [
-        compileSdk : 30,
-        minSdk     : 21,
+        compileSdk : 31,
+        minSdk     : 26,
         targetSdk  : 30,
         versionCode: 1,
         versionName: "1.0",
-        'androidTools': '30.0.0',
+        'androidTools': '31.0.0',
         'kotlin': '1.7.10',
-        'incap' : '0.2',
 
-        'release': '8.8.1'
+        'butterknife': '10.2.3'
     ]
+
     ext.deps = [
         'androidx'    : [
             'appcompat'         : "androidx.appcompat:appcompat:1.3.1",
             'constraintlayout'  : "androidx.constraintlayout:constraintlayout:2.1.0",
-            'core'              : "androidx.core:core:1.0.0",
-            'viewpager'         : "androidx.viewpager:viewpager:1.0.0",
-            'annotations'       : "androidx.annotation:annotation:1.0.0",
+            'core'              : "androidx.core:core:1.3.0",
+            'corektx'          : "androidx.core:core-ktx:1.7.0",
             'test'              : [
                 'junit'         : 'androidx.test.ext:junit:1.1.3',
                 'espressoCore' : 'androidx.test.espresso:espresso-core:3.4.0',
-                'runner'        : 'androidx.test:runner:1.1.1',
-                'rules'         : 'androidx.test:rules:1.1.1',
+                'runner'        : 'androidx.test:runner:1.3.1',
             ],
         ],
-        material      : 'com.google.android.material:material:1.4.0',
-        material      : 'com.google.android.material:material:1.4.0',
-        'lint'        : [
-            'core'  : "com.android.tools.lint:lint:${versions.androidTools}",
-            'api'   : "com.android.tools.lint:lint-api:${versions.androidTools}",
-            'checks': "com.android.tools.lint:lint-checks:${versions.androidTools}",
-            'tests' : "com.android.tools.lint:lint-tests:${versions.androidTools}",
-        ],
 
-        javapoet      : 'com.squareup:javapoet:1.10.0',
+        material      : 'com.google.android.material:material:1.4.0',
+
+        javapoet      : 'com.squareup:javapoet:1.13.0',
         junit         : 'junit:junit:4.13.2',
-        truth         : 'com.google.truth:truth:0.42',
-        compiletesting: 'com.google.testing.compile:compile-testing:0.15',
         'auto'        : [
             'service': 'com.google.auto.service:auto-service:1.0-rc4',
             'common' : 'com.google.auto:auto-common:0.10',
         ],
-        'guava'       : 'com.google.guava:guava:24.0-jre',
+
         'release'     : [
-            'runtime' : "com.jakewharton:butterknife:${versions.release}",
-            'com.butterknife.compiler': "com.jakewharton:butterknife-com.butterknife.compiler:${versions.release}"
+            'runtime' : "com.jakewharton:butterknife:${versions.butterknife}",
+            'com.butterknife.compiler': "com.jakewharton:butterknife-com.butterknife.compiler:${versions.butterknife}"
         ],
         'kotlin'      : [
             'stdLibJdk8': "org.jetbrains.kotlin:kotlin-stdlib-jdk8:${versions.kotlin}",
-        ],
-        'incap'       : [
-            'runtime'  : "net.ltgt.gradle.incap:incap:${versions.incap}",
-            'processor': "net.ltgt.gradle.incap:incap-processor:${versions.incap}",
         ]
     ]
     repositories {
@@ -62,7 +48,6 @@ buildscript {
         google()
     }
     dependencies {
-//        classpath "com.android.tools.build:gradle:4.0.1"
         classpath 'com.android.tools.build:gradle:7.2.1'
         classpath 'org.jetbrains.kotlin:kotlin-gradle-plugin:1.7.10'
         // NOTE: Do not place your application dependencies here; they belong
